BotTopicQuestionAsked.py
DOES NOT CURRENTLY WORK, NEEDS DatabaseTopicPrompt FIRST

Created BotTopicQuestionAsked.py, which contains a class/function that take in a list and compare said list to the database prompts list.

If the lists are the same size, empty the inputted list and return a random prompt from the database.

If the lists are not the same size, keep getting a new prompt from the database until the prompt is not one that is in the inputted list and return it.