BotSentimentResponse.py
CURRENTLY DOES NOT WORK, NEEDS BroadQuestion.py AND DatabaseTopicPrompt.py IMPLEMENTED TO WORK

BotSentimentResponse.py takes in a string and a list. It then gets the sentence sentiment of the string, and checks it.

If the sentiment is >0, returns a random string from BroadQuestion.py

Else, gets a new string from BotTopicQuestionAsked.py, appends it to the inputted string, and then returns the new string.